plat-sam: move pl310 related code to its own file
Cleanup main.c by moving pl310 code to sam_pl310.c file.

Acked-by: Jerome Forissier <jerome@forissier.org>
Acked-by: Etienne Carriere <etienne.carriere@linaro.org>
