+++ core/arch/arm/plat-sam/main.c
@@ -34,12 +34,12 @@
 #include <kernel/misc.h>
 #include <kernel/panic.h>
 #include <kernel/tz_ssvce_def.h>
-#include <kernel/tz_ssvce_pl310.h>
 #include <matrix.h>
 #include <mm/core_mmu.h>
 #include <mm/core_memprot.h>
 #include <platform_config.h>
 #include <sama5d2.h>
+#include <sam_sfr.h>
 #include <stdint.h>
 #include <sm/optee_smc.h>
 #include <tz_matrix.h>
@@ -54,10 +54,9 @@ void console_init(void)
 	register_serial_console(&console_data.chip);
 }
 
-register_phys_mem_pgdir(MEM_AREA_IO_SEC, PL310_BASE, CORE_MMU_PGDIR_SIZE);
 register_phys_mem_pgdir(MEM_AREA_IO_SEC, SFR_BASE, CORE_MMU_PGDIR_SIZE);
 
-static vaddr_t sfr_base(void)
+vaddr_t sam_sfr_base(void)
 {
 	static void *va;
 
@@ -67,46 +66,6 @@ static vaddr_t sfr_base(void)
 		return (vaddr_t)va;
 	}
 	return SFR_BASE;
-}
-
-enum ram_config {RAMC_SRAM = 0, RAMC_L2CC};
-
-static void l2_sram_config(enum ram_config setting)
-{
-	if (setting == RAMC_L2CC)
-		io_write32(sfr_base() + SFR_L2CC_HRAMC, 0x1);
-	else
-		io_write32(sfr_base() + SFR_L2CC_HRAMC, 0x0);
-}
-
-vaddr_t pl310_base(void)
-{
-	static void *va;
-
-	if (cpu_mmu_enabled()) {
-		if (!va)
-			va = phys_to_virt(PL310_BASE, MEM_AREA_IO_SEC, 1);
-		return (vaddr_t)va;
-	}
-	return PL310_BASE;
-}
-
-void arm_cl2_config(vaddr_t pl310_base)
-{
-	io_write32(pl310_base + PL310_CTRL, 0);
-	l2_sram_config(RAMC_L2CC);
-	io_write32(pl310_base + PL310_AUX_CTRL, PL310_AUX_CTRL_INIT);
-	io_write32(pl310_base + PL310_PREFETCH_CTRL, PL310_PREFETCH_CTRL_INIT);
-	io_write32(pl310_base + PL310_POWER_CTRL, PL310_POWER_CTRL_INIT);
-
-	/* invalidate all cache ways */
-	arm_cl2_invbyway(pl310_base);
-}
-
-void arm_cl2_enable(vaddr_t pl310_base)
-{
-	/* Enable PL310 ctrl -> only set lsb bit */
-	io_write32(pl310_base + PL310_CTRL, 1);
 }
 
 register_phys_mem_pgdir(MEM_AREA_IO_SEC, AT91C_BASE_MATRIX32,
